refactors : login refactoring
