+++ newsApp/app.js
@@ -7,13 +7,7 @@ const session = require('express-session');
 const flash = require('connect-flash');
 require('dotenv').config();
 
-// const pageRouter = require('./routes/index');
-const indexRouter = require('./routes/index');
-const authRouter = require('./routes/auth');
-// const connect = require('./schemas')
-
 const app = express();
-// connect()
 
 mongoose.connect('mongodb://localhost:27017/myapp', {useNewUrlParser: true})
 
@@ -21,12 +15,7 @@ app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'pug');
 app.set('port', process.env.PORT || 8001);
 
-// app.use(function(req, res, next) {
-//   res.render('index');
-// });
-
 app.use(morgan('dev'));
-// app.use(express.static(path.join(__dirname, 'public')));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser(process.env.COOKIE_SECRET));
@@ -41,53 +30,9 @@ app.use(session({
 }));
 app.use(flash());
 
-const authData = {
-  email : "123",
-  nick : "ruby",
-  password : "123"
-}
-var passport = require('passport')
-  , LocalStrategy = require('passport-local').Strategy;
-
-app.use(passport.initialize())
-app.use(passport.session())
-
-passport.serializeUser(function(user, done) {
-  console.log("serializeUser",user)
-  done(null, user.email);
-});
-
-passport.deserializeUser(function(id, done) {
-  console.log("deserializeUser",id)
-  done(null, authData);
-});
-
-passport.use(new LocalStrategy(
-  {
-    usernameField : 'email',
-    passwordField : 'password'
-  },
-  function(email, password, done) {
-    if(email === authData.email){
-      if(password === authData.password){
-        return done(null, authData);
-      }else{
-        return done(null, false, { message: 'Incorrect password.' });
-      }
-    }else{
-      return done(null, false, { message: 'Incorrect username.' });
-    }
-  }
-));
-
-
-app.post('/auth/login',
-  passport.authenticate('local', { 
-    successRedirect: '/',
-    // failureRedirect: '/login',
-    failureFlash : true
-  })
-);
+const passport = require('./passport/index')(app)
+const indexRouter = require('./routes/index');
+const authRouter = require('./routes/auth')(passport);
 
 app.use('/', indexRouter);
 app.use('/auth', authRouter);
