--- newsApp/test/test.js
@@ -1,3 +1,0 @@
-test('1 is 1', () => {
-  expect(1).toBe(1);
-});
